refactor(quickstart): split endpoint and usage helpers
